to modify events and show images

+++ frontend/src/admin/Dashboard.jsx
@@ -1,16 +1,19 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
-import './components/dashboard.css'
+import './components/style.css'
 import { Link } from 'react-router-dom';
 import Header from './components/Header';
 import Sidebar from './components/Sidebar';
-import Footer from './components//Footer';
+import Footer from './components/Footer';
+import Swal from 'sweetalert2';
 
 const Dashboard = () => {
     const [data, setData] = useState([]);
     const [modalVisible, setModalVisible] = useState(false);
     const [selectedData, setSelectedData] = useState({});
     const [sortDirection, setSortDirection] = useState({ key: '', direction: 'asc' });
+    const [toggleArrow, setToggleArrow] = useState(false);
+
 
     useEffect(() => {
         document.title = "Dashboard";
@@ -47,6 +50,7 @@ const Dashboard = () => {
     };
 
     const handleSort = (key) => {
+        // alert(typeof key)
         const direction = sortDirection.key === key && sortDirection.direction === 'asc' ? 'desc' : 'asc';
         setSortDirection({ key, direction });
         const sortedData = [...data].sort((a, b) => {
@@ -54,6 +58,10 @@ const Dashboard = () => {
             if (a[key] > b[key]) return direction === 'asc' ? 1 : -1;
             return 0;
         });
+        setToggleArrow (true)
+        if(toggleArrow === true){
+            setToggleArrow (false)
+        }
         setData(sortedData);
     };
     const handlechange = async (e) => {
@@ -71,14 +79,29 @@ const Dashboard = () => {
     const Delete=async(id)=>{
         try {
             // alert(id)
-            const response=await axios.delete(`http://localhost:8081/api/deleteEvent/${id}`,{headers:{Authorization: `Bearer ${localStorage.getItem('token')}`}})
-            // console.log(response.data)
-            // setData(response.data);
-            fetchAllData()
-        } catch (error) {
+            const confirmed=await Swal.fire({
+                title: "Are you sure?",
+                text: "You won't be able to revert this!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Yes, delete it!"
+              });
+              if(confirmed.isConfirmed){
+                const response=await axios.delete(`http://localhost:8081/api/deleteEvent/${id}`,{headers:{Authorization: `Bearer ${localStorage.getItem('token')}`}})
+                // console.log(response.data)
+                // setData(response.data);
+                fetchAllData()
+              }else{
+                alert("you cancelled!")
+              }
             
+        } catch (error) {
+            console.log(error)
         }
     }
+    
 
     return (
         <div className="dashboard-layout">
@@ -100,24 +123,27 @@ const Dashboard = () => {
                         </div>
                         <table>
                             <thead>
+           
                                 <tr>
-                                    <th onClick={() => handleSort('id')}>Sl No</th>
-                                    <th onClick={() => handleSort('title')}>Title</th>
-                                    <th>Description</th>
-                                    <th>Date</th>
-                                    <th>Time</th>
-                                    <th>Location</th>
-                                    <th>General Ticket Price</th>
-                                    <th>Vip Ticket Price</th>
-                                    <th>Quantity</th>
-                                    <th>Actions</th>
+                                    
+                                    <th onClick={() => handleSort('id')} style={{width:"61px"}}>Sl No {toggleArrow === false ?  <span class="fa fa-arrow-up" style={{color:'rgb(91, 245, 2)'}}></span> : <span class="fa fa-arrow-down" style={{color:'rgb(245, 2, 34)'}}></span>}</th>
+                                    <th onClick={() => handleSort('title')} style={{width:"120px"}}>Title {toggleArrow === false ?  <span class="fa fa-arrow-up" style={{color:'rgb(91, 245, 2)'}}></span> : <span class="fa fa-arrow-down" style={{color:'rgb(245, 2, 34)'}}></span>}</th>
+                                    <th style={{width:"161px"}}>Description</th>
+                                    <th >Date</th>
+                                    <th style={{width:"61px"}}>Time</th>
+                                    <th style={{width:"61px"}}>Location</th>
+                                    <th >General Ticket Price</th>
+                                    <th >Vip Ticket Price</th>
+                                    <th >Quantity</th>
+                                    <th style={{width:'95px'}}>Actions</th>
                                 </tr>
                             </thead>
                             <tbody>
                                 {data.map((event, index) => (
                                     <tr key={index}>
-                                        <td>{index+1}</td>
-                                        <td>{event.title}</td>
+                                        
+                                        <td style={{width:"15px"}}>{index+1}</td>
+                                        <td style={{width:"25px"}}>{event.title}</td>
                                         <td>{event.description}</td>
                                         <td>{event.eventDate}</td>
                                         <td>{event.eventTime}</td>
@@ -125,11 +151,10 @@ const Dashboard = () => {
                                         <td>{event.general_ticket_price}</td>
                                         <td>{event.vip_ticket_price}</td>
                                         <td>{event.quantity}</td>
-                                        <td>
-                                            <button className="edit" style={{background:"green"}}><Link to={`/addedit/${event.id}`} style={{ textDecoration: 'none', color: 'white' }}>Edit</Link></button>
-                                            <button onClick={() => viewData(event)} style={{background:"black", color:"white"}}>View</button>
-                                            <button onClick={() => Delete(event.id)} style={{background:"red", color:"white"}}>Del</button>
-                                            
+                                        <td style={{padding:'0px'}}>
+                                            <button className="edit" style={{background:"green"}}><Link to={`/addedit/${event.id}`} style={{ textDecoration: 'none', color: 'white' }}><i class="fa fa-edit"></i></Link></button>
+                                            <button onClick={() => viewData(event)} style={{background:"black", color:"white", margin:"5px"}}><i class="fa fa-eye" aria-hidden="true"></i></button>
+                                            <button onClick={() => Delete(event.id)} style={{background:"red", color:"white"}}><i class="fa fa-trash" aria-hidden="true"></i></button>                                            
                                         </td>
                                     </tr>
                                 ))}
